@@ -5,75 +5,74 @@
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
-    "start": "yarn build-themes && ng serve",
-    "start:jit": "yarn build-themes && ng serve --aot=false",
-    "start:prod": "yarn build-themes && ng serve --prod",
+    "start": "ng serve",
+    "start:jit": "ng serve --aot=false",
+    "start:prod": "ng serve --prod",
+    "start:scenes": "ng serve scenes",
     "lint": "ng lint",
-    "test": "yarn build-themes && ng test",
+    "test": "ng test",
     "pree2e": "webdriver-manager update",
-    "e2e": "yarn build-themes && ng e2e",
-    "build": "yarn build-themes && ng build",
+    "e2e": "ng e2e",
+    "generate-scenes": "ng e2e --port 4201 scenes",
+    "build": "ng build",
+    "build:scenes": "ng build scenes",
     "build:content": "yarn upgrade @angular/components-examples",
-    "build:highmem": "yarn build-themes && node --max_old_space_size=8192 ./node_modules/@angular/cli/bin/ng build",
     "build:sm": "ng build --prod --source-map",
-    "build-themes": "bash ./tools/build-themes.sh",
-    "prod-build": "yarn build-themes && ng build --prod",
-    "postinstall": "webdriver-manager update --gecko false && yarn build-themes && ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
+    "prod-build": "ng build --prod",
+    "postinstall": "webdriver-manager update --gecko false && ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
     "publish-prod": "bash ./tools/deploy.sh stable prod",
     "publish-dev": "bash ./tools/deploy.sh",
     "publish-beta": "bash ./tools/deploy.sh stable beta"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.0.0",
-    "@angular/cdk": "^9.0.1",
-    "@angular/cdk-experimental": "^9.0.1",
-    "@angular/common": "9.0.0",
-    "@angular/compiler": "9.0.0",
-    "@angular/components-examples": "angular/material2-docs-content#9.0.x",
-    "@angular/core": "9.0.0",
-    "@angular/forms": "9.0.0",
-    "@angular/google-maps": "^9.0.1",
-    "@angular/material": "^9.0.1",
-    "@angular/material-experimental": "^9.0.1",
-    "@angular/material-moment-adapter": "^9.0.1",
-    "@angular/platform-browser": "9.0.0",
-    "@angular/platform-browser-dynamic": "9.0.0",
-    "@angular/router": "9.0.0",
-    "@angular/youtube-player": "^9.0.1",
-    "core-js": "^2.6.11",
-    "material-components-web": "^4.0.0",
-    "moment": "^2.24.0",
-    "rxjs": "^6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "^11.0.0",
+    "@angular/cdk": "^11.0.3",
+    "@angular/cdk-experimental": "^11.0.3",
+    "@angular/common": "^11.0.0",
+    "@angular/compiler": "^11.0.0",
+    "@angular/components-examples": "angular/material2-docs-content#11.0.x",
+    "@angular/core": "^11.0.0",
+    "@angular/forms": "^11.0.0",
+    "@angular/google-maps": "^11.0.3",
+    "@angular/material": "^11.0.3",
+    "@angular/material-experimental": "^11.0.3",
+    "@angular/material-moment-adapter": "^11.0.3",
+    "@angular/platform-browser": "^11.0.0",
+    "@angular/platform-browser-dynamic": "^11.0.0",
+    "@angular/router": "^11.0.0",
+    "@angular/youtube-player": "^11.0.3",
+    "material-components-web": "9.0.0-canary.419e03572.0",
+    "moment": "^2.29.1",
+    "rxjs": "^6.6.3",
+    "tslib": "^2.0.3",
+    "zone.js": "^0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.900.0",
-    "@angular/cli": "9.0.0",
-    "@angular/compiler-cli": "9.0.0",
-    "@angular/localize": "9.0.0",
+    "@angular-devkit/build-angular": "0.1100.0",
+    "@angular/cli": "11.0.0",
+    "@angular/compiler-cli": "^11.0.0",
+    "@angular/localize": "^11.0.0",
     "@awesome-fe/translate": "^0.3.2",
-    "@types/jasmine": "^3.5.2",
-    "@types/node": "^12.12.24",
-    "codelyzer": "^5.2.1",
-    "firebase-tools": "^7.12.1",
+    "@types/jasmine": "^3.6.0",
+    "@types/node": "^12.12.70",
+    "codelyzer": "^6.0.1",
+    "firebase-tools": "^8.16.0",
     "gulp": "^3.9.1",
     "gulp-replace": "^1.0.0",
     "hammerjs": "^2.0.8",
     "highlight.js": "^9.9.0",
-    "jasmine-core": "^3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.4.1",
+    "jasmine-core": "^3.6.0",
+    "jasmine-spec-reporter": "^5.0.2",
+    "karma": "~5.1.1",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-firefox-launcher": "^1.3.0",
-    "karma-jasmine": "^3.1.0",
-    "karma-jasmine-html-reporter": "^1.5.1",
-    "node-sass": "^4.13.1",
-    "protractor": "^5.4.3",
-    "ts-node": "8.6.2",
-    "tslint": "^5.20.1",
-    "typescript": "3.6.4"
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-firefox-launcher": "^2.1.0",
+    "karma-jasmine": "^4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "^7.0.0",
+    "ts-node": "^8.10.1",
+    "tslint": "^6.1.2",
+    "typescript": "4.0.2"
   }
 }
