@@ -11,35 +11,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.0.0",
-    "@angular/cdk": "^9.0.0",
-    "@angular/common": "^9.0.0",
-    "@angular/compiler": "^9.0.0",
-    "@angular/core": "^9.0.0",
-    "@angular/forms": "^9.0.0",
-    "@angular/material": "^9.0.0",
-    "@angular/material-moment-adapter": "^9.0.0",
-    "@angular/platform-browser": "^9.0.0",
-    "@angular/platform-browser-dynamic": "^9.0.0",
-    "@angular/router": "^9.0.0",
-    "angular-in-memory-web-api": "~0.9.0",
-    "core-js": "^2.6.9",
-    "hammerjs": "^2.0.8",
-    "moment": "^2.24.0",
-    "rxjs": "^6.5.3",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "^11.0.0",
+    "@angular/cdk": "^11.0.0",
+    "@angular/common": "^11.0.0",
+    "@angular/compiler": "^11.0.0",
+    "@angular/core": "^11.0.0",
+    "@angular/forms": "^11.0.0",
+    "@angular/material": "^11.0.0",
+    "@angular/material-moment-adapter": "^11.0.0",
+    "@angular/platform-browser": "^11.0.0",
+    "@angular/platform-browser-dynamic": "^11.0.0",
+    "@angular/router": "^11.0.0",
+    "angular-in-memory-web-api": "~0.11.0",
+    "moment": "^2.29.1",
+    "rxjs": "^6.6.3",
+    "tslib": "^2.0.3",
+    "zone.js": "^0.11.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.900.0",
-    "@angular/cli": "^9.0.0",
-    "@angular/compiler-cli": "^9.0.0",
-    "@angular/language-service": "^9.0.0",
-    "@angular/localize": "^9.0.0",
-    "@types/node": "^12.12.6",
-    "codelyzer": "^5.2.0",
-    "ts-node": "^8.4.1",
-    "tslint": "^5.20.1",
-    "typescript": "3.7.5"
+    "@angular-devkit/build-angular": "0.1100.0",
+    "@angular/cli": "^11.0.0",
+    "@angular/compiler-cli": "^11.0.0",
+    "@angular/language-service": "^11.0.0",
+    "@angular/localize": "^11.0.0",
+    "@types/node": "^12.12.70",
+    "codelyzer": "^6.0.1",
+    "ts-node": "^8.10.1",
+    "tslint": "~6.1.2",
+    "typescript": "~4.0.2"
   }
 }
