@@ -1,8 +1,9 @@
-import {Component, ViewEncapsulation} from '@angular/core';
-import {Event, NavigationEnd, Router} from '@angular/router';
-import {filter} from 'rxjs/operators';
+import {Component, OnDestroy, ViewEncapsulation} from '@angular/core';
 
 import {GaService} from './shared/ga/ga';
+import {NavigationFocusService} from './shared/navigation-focus/navigation-focus.service';
+import {Subscription} from 'rxjs';
+import {map, pairwise, startWith} from 'rxjs/operators';
 
 @Component({
   selector: 'material-docs-app',
@@ -10,36 +11,27 @@ import {GaService} from './shared/ga/ga';
   styleUrls: ['./material-docs-app.scss'],
   encapsulation: ViewEncapsulation.None,
 })
-export class MaterialDocsApp {
+export class MaterialDocsApp implements OnDestroy {
+  private subscriptions = new Subscription();
 
-  constructor(router: Router, ga: GaService) {
-    let previousRoute = router.routerState.snapshot.url;
-
-    router.events
-      .pipe(filter((event: Event) => event instanceof NavigationEnd))
-      .subscribe((data: Event) => {
-        const urlAfterRedirects = (data as NavigationEnd).urlAfterRedirects;
-        // We want to reset the scroll position on navigation except when navigating within
-        // the documentation for a single component.
-        if (!isNavigationWithinComponentView(previousRoute, urlAfterRedirects)) {
-          resetScrollPosition();
-        }
-
-        previousRoute = urlAfterRedirects;
-        ga.locationChanged(urlAfterRedirects);
-      });
+  constructor(ga: GaService, navigationFocusService: NavigationFocusService) {
+    this.subscriptions.add(navigationFocusService.navigationEndEvents.pipe(
+      map(e => e.urlAfterRedirects),
+      startWith(''),
+      pairwise()
+    ).subscribe(([fromUrl, toUrl]) => {
+      // We want to reset the scroll position on navigation except when navigating within
+      // the documentation for a single component.
+      if (!navigationFocusService.isNavigationWithinComponentView(fromUrl, toUrl)) {
+        resetScrollPosition();
+      }
+      ga.locationChanged(toUrl);
+    }));
   }
-}
 
-function isNavigationWithinComponentView(previousUrl: string, newUrl: string) {
-  const componentViewExpression = /(cdk|components)\/(\w+)/;
-
-  const previousUrlMatch = previousUrl.match(componentViewExpression);
-  const newUrlMatch = newUrl.match(componentViewExpression);
-
-  return previousUrl && newUrl && previousUrlMatch && newUrlMatch
-      && previousUrlMatch[0] === newUrlMatch[0]
-      && previousUrlMatch[1] === newUrlMatch[1];
+  ngOnDestroy() {
+    this.subscriptions.unsubscribe();
+  }
 }
 
 function resetScrollPosition() {
