@@ -4,7 +4,9 @@ import {MatListModule} from '@angular/material/list';
 import {RouterModule, Routes} from '@angular/router';
 import {FooterModule} from '../../shared/footer/footer';
 import {CommonModule} from '@angular/common';
+import {NavigationFocusModule} from '../../shared/navigation-focus/navigation-focus';
 import {ComponentPageTitle} from '../page-title/page-title';
+import {MatCardModule} from '@angular/material/card';
 
 @Component({
   selector: 'app-guides',
@@ -19,10 +21,14 @@ export class GuideList implements OnInit {
   }
 }
 
-const routes: Routes = [ {path : '', component : GuideList} ];
+const routes: Routes = [{path: '', component: GuideList}];
 
 @NgModule({
-  imports: [CommonModule, MatListModule, FooterModule, RouterModule.forChild(routes)],
+  imports: [CommonModule,
+            MatListModule,
+            FooterModule,
+            RouterModule.forChild(routes),
+            NavigationFocusModule, MatCardModule],
   exports: [GuideList],
   declarations: [GuideList],
   providers: [GuideItems],
