@@ -10,7 +10,7 @@ import {PortalModule} from '@angular/cdk/portal';
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {HeaderLink} from './header-link';
-import {CopierService} from '../copier/copier.service';
+import {CodeSnippet} from '../example-viewer/code-snippet';
 
 
 // ExampleViewer is included in the DocViewerModule because they have a circular dependency.
@@ -25,8 +25,7 @@ import {CopierService} from '../copier/copier.service';
     PortalModule,
     StackBlitzButtonModule
   ],
-  providers: [CopierService],
-  declarations: [DocViewer, ExampleViewer, HeaderLink],
+  declarations: [DocViewer, ExampleViewer, HeaderLink, CodeSnippet],
   entryComponents: [ExampleViewer, HeaderLink],
   exports: [DocViewer, ExampleViewer, HeaderLink],
 })
