@@ -1,16 +1,17 @@
 import {HttpTestingController} from '@angular/common/http/testing';
 import {Component} from '@angular/core';
-import {async, inject, TestBed} from '@angular/core/testing';
+import {waitForAsync, inject, TestBed} from '@angular/core/testing';
 import {By} from '@angular/platform-browser';
 import {DocsAppTestingModule} from '../../testing/testing-module';
 import {DocViewer} from './doc-viewer';
 import {DocViewerModule} from './doc-viewer-module';
+import {ExampleViewer} from '../example-viewer/example-viewer';
 
 
 describe('DocViewer', () => {
   let http: HttpTestingController;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       imports: [DocViewerModule, DocsAppTestingModule],
       declarations: [DocViewerTestComponent],
@@ -70,6 +71,52 @@ describe('DocViewer', () => {
     expect(docViewer.nativeElement.innerHTML).toContain('id="my-header"');
   });
 
+  it('should instantiate example viewer in snippet view with region', () => {
+    const fixture = TestBed.createComponent(DocViewerTestComponent);
+    const testUrl = 'http://material.angular.io/snippet-example.html';
+
+    fixture.componentInstance.documentUrl = testUrl;
+    fixture.detectChanges();
+
+    http.expectOne(testUrl).flush(FAKE_DOCS[testUrl]);
+
+    const exampleViewer = fixture.debugElement.query(By.directive(ExampleViewer));
+    expect(exampleViewer.componentInstance.file).toBe('some-example.html');
+    expect(exampleViewer.componentInstance.showCompactToggle).toBeTrue();
+    expect(exampleViewer.componentInstance.region).toBe('some-region');
+    expect(exampleViewer.componentInstance.view).toBe('snippet');
+  });
+
+  it('should instantiate example viewer in demo view', () => {
+    const fixture = TestBed.createComponent(DocViewerTestComponent);
+    const testUrl = 'http://material.angular.io/demo-example.html';
+
+    fixture.componentInstance.documentUrl = testUrl;
+    fixture.detectChanges();
+
+    http.expectOne(testUrl).flush(FAKE_DOCS[testUrl]);
+
+    const exampleViewer = fixture.debugElement.query(By.directive(ExampleViewer));
+    expect(exampleViewer.componentInstance.file).toBeUndefined();
+    expect(exampleViewer.componentInstance.showCompactToggle).toBeFalse();
+    expect(exampleViewer.componentInstance.view).toBe('demo');
+  });
+
+  it('should instantiate example viewer in snippet view with whole snippet', () => {
+    const fixture = TestBed.createComponent(DocViewerTestComponent);
+    const testUrl = 'http://material.angular.io/whole-snippet-example.html';
+
+    fixture.componentInstance.documentUrl = testUrl;
+    fixture.detectChanges();
+
+    http.expectOne(testUrl).flush(FAKE_DOCS[testUrl]);
+
+    const exampleViewer = fixture.debugElement.query(By.directive(ExampleViewer));
+    expect(exampleViewer.componentInstance.file).toBe('whole-snippet-example.ts');
+    expect(exampleViewer.componentInstance.showCompactToggle).toBeTrue();
+    expect(exampleViewer.componentInstance.view).toBe('snippet');
+  });
+
   it('should show error message when doc not found', () => {
     spyOn(console, 'log');
 
@@ -112,4 +159,11 @@ const FAKE_DOCS: {[key: string]: string} = {
       <div material-docs-example="some-example"></div>`,
   'http://material.angular.io/doc-with-links.html': `<a href="#test">Test link</a>`,
   'http://material.angular.io/doc-with-element-ids.html': `<h4 id="my-header">Header</h4>`,
+  'http://material.angular.io/snippet-example.html':
+    '<div material-docs-example="some-example" file="some-example.html"' +
+    ' region="some-region"></div>',
+  'http://material.angular.io/demo-example.html':
+    '<div material-docs-example="demo-example"></div>',
+  'http://material.angular.io/whole-snippet-example.html':
+    '<div material-docs-example="whole-snippet-example" file="whole-snippet-example.ts"></div>',
 };
