@@ -1,31 +1,36 @@
 import {CommonModule} from '@angular/common';
 import {HttpTestingController} from '@angular/common/http/testing';
 import {NgModule} from '@angular/core';
-import {async, ComponentFixture, inject, TestBed} from '@angular/core/testing';
+import {waitForAsync, ComponentFixture, inject, TestBed} from '@angular/core/testing';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {MatAutocompleteModule} from '@angular/material/autocomplete';
 import {MatInputModule} from '@angular/material/input';
 import {MatSlideToggleModule} from '@angular/material/slide-toggle';
 import {MatSnackBar} from '@angular/material/snack-bar';
+import {Clipboard} from '@angular/cdk/clipboard';
 
 import {EXAMPLE_COMPONENTS} from '@angular/components-examples';
 import {By} from '@angular/platform-browser';
 import {NoopAnimationsModule} from '@angular/platform-browser/animations';
 import {DocsAppTestingModule} from '../../testing/testing-module';
-import {CopierService} from '../copier/copier.service';
 import {DocViewerModule} from '../doc-viewer/doc-viewer-module';
 import {ExampleViewer} from './example-viewer';
 import {AutocompleteExamplesModule} from '@angular/components-examples/material/autocomplete';
+import {MatTabGroupHarness} from '@angular/material/tabs/testing';
+import {HarnessLoader} from '@angular/cdk/testing';
+import {TestbedHarnessEnvironment} from '@angular/cdk/testing/testbed';
 
 const exampleKey = 'autocomplete-overview';
+const exampleBasePath = `/docs-content/examples-highlighted/material/autocomplete/${exampleKey}`;
 
 
 describe('ExampleViewer', () => {
   let fixture: ComponentFixture<ExampleViewer>;
   let component: ExampleViewer;
   let http: HttpTestingController;
+  let loader: HarnessLoader;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       imports: [
         DocViewerModule,
@@ -43,42 +48,100 @@ describe('ExampleViewer', () => {
   beforeEach(() => {
     fixture = TestBed.createComponent(ExampleViewer);
     component = fixture.componentInstance;
+    fixture.detectChanges();
+    loader = TestbedHarnessEnvironment.loader(fixture);
   });
 
-  it('should toggle showSource boolean', async(() => {
-    fixture.detectChanges();
-    expect(component.showSource).toBe(false);
+  it('should toggle between the 3 views', waitForAsync(() => {
+    // need to specify a file because toggling from snippet to full changes the tabs to match
+    component.file = 'file.html';
+    component.view = 'snippet';
+    expect(component.view).toBe('snippet');
+    component.toggleCompactView();
+    expect(component.view).toBe('full');
     component.toggleSourceView();
-    expect(component.showSource).toBe(true);
+    expect(component.view).toBe('demo');
   }));
 
-  it('should set and return example properly', async(() => {
+  it('should expand to HTML tab', waitForAsync(async () => {
     component.example = exampleKey;
-    fixture.detectChanges();
-    const data = component.exampleData;
-    // TODO(jelbourn): remove `as any` once LiveExample is updated to have optional members.
-    expect(data).toEqual(EXAMPLE_COMPONENTS[exampleKey] as any);
+    component.file = 'file.html';
+    component.view = 'snippet';
+    component.toggleCompactView();
+
+    const tabGroup = await loader.getHarness(MatTabGroupHarness);
+    const tab = await tabGroup.getSelectedTab();
+    expect(await tab.getLabel()).toBe('HTML');
   }));
 
-  it('should print an error message about missing example', async(() => {
+  it('should expand to TS tab', waitForAsync(async () => {
+    component.example = exampleKey;
+    component.file = EXAMPLE_COMPONENTS[exampleKey].primaryFile;
+    component.view = 'snippet';
+    component.toggleCompactView();
+
+    const tabGroup = await loader.getHarness(MatTabGroupHarness);
+    const tab = await tabGroup.getSelectedTab();
+    expect(await tab.getLabel()).toBe('TS');
+  }));
+
+  it('should expand to CSS tab', waitForAsync(async () => {
+    component.example = exampleKey;
+    component.file = 'file.css';
+    component.view = 'snippet';
+    component.toggleCompactView();
+
+    const tabGroup = await loader.getHarness(MatTabGroupHarness);
+    const tab = await tabGroup.getSelectedTab();
+    expect(await tab.getLabel()).toBe('CSS');
+  }));
+
+  it('should generate correct url with region', waitForAsync(() => {
+    component.example = exampleKey;
+    component.region = 'region';
+    const url = component.generateUrl('a.b.html');
+    expect(url).toBe(exampleBasePath + '/a.b_region-html.html');
+  }));
+
+  it('should generate correct url without region', waitForAsync(() => {
+    component.example = exampleKey;
+    component.region = undefined;
+    const url = component.generateUrl('a.b.ts');
+    expect(url).toBe(exampleBasePath + '/a.b-ts.html');
+  }));
+
+  it('should print an error message about incorrect file type', waitForAsync(() => {
+    spyOn(console, 'error');
+    component.example = exampleKey;
+    component.file = 'file.bad';
+    component.selectCorrectTab();
+
+    expect(console.error).toHaveBeenCalledWith(
+      `Could not find tab for file extension: "bad".`);
+  }));
+
+  it('should set and return example properly', waitForAsync(() => {
+    component.example = exampleKey;
+    const data = component.exampleData;
+    expect(data).toEqual(EXAMPLE_COMPONENTS[exampleKey]);
+  }));
+
+  it('should print an error message about missing example', waitForAsync(() => {
     spyOn(console, 'error');
     component.example = 'foobar';
-    fixture.detectChanges();
     expect(console.error).toHaveBeenCalled();
     expect(console.error).toHaveBeenCalledWith('Could not find example: foobar');
   }));
 
-  it('should return docs-content path for example based on extension', async(() => {
+  it('should return docs-content path for example based on extension', waitForAsync(() => {
     // set example
     component.example = exampleKey;
-    fixture.detectChanges();
 
     // get example file path for each extension
     const extensions = ['ts', 'css', 'html'];
-    const basePath = '/docs-content/examples-highlighted/';
 
     extensions.forEach(extension => {
-      const expected = `${basePath}${exampleKey}-example-${extension}.html`;
+      const expected = `${exampleBasePath}/${exampleKey}-example-${extension}.html`;
       const actual = component.exampleTabs[extension.toUpperCase()];
 
       expect(actual).toEqual(expected);
@@ -89,7 +152,6 @@ describe('ExampleViewer', () => {
 
     it('should only render HTML, TS and CSS files if no additional files are specified', () => {
       component.example = exampleKey;
-      fixture.detectChanges();
 
       expect(component._getExampleTabNames()).toEqual(['HTML', 'TS', 'CSS']);
     });
@@ -97,14 +159,31 @@ describe('ExampleViewer', () => {
     it('should be able to render additional files', () => {
       EXAMPLE_COMPONENTS['additional-files'] = {
         ...EXAMPLE_COMPONENTS[exampleKey],
-        additionalFiles: ['some-additional-file.html'],
+        files: [
+          'additional-files-example.ts',
+          'additional-files-example.html',
+          'additional-files-example.css',
+          'some-additional-file.html'
+        ],
       };
 
       component.example = 'additional-files';
-      fixture.detectChanges();
 
       expect(component._getExampleTabNames())
         .toEqual(['HTML', 'TS', 'CSS', 'some-additional-file.html']);
+    });
+
+    it('should be possible for example to not have CSS or HTML files', () => {
+      EXAMPLE_COMPONENTS['additional-files'] = {
+        ...EXAMPLE_COMPONENTS[exampleKey],
+        files: [
+          'additional-files-example.ts',
+        ],
+      };
+
+      component.example = 'additional-files';
+
+      expect(component._getExampleTabNames()).toEqual(['TS']);
     });
   });
 
@@ -114,7 +193,7 @@ describe('ExampleViewer', () => {
     beforeEach(() => {
       // Open source view
       component.example = exampleKey;
-      component.showSource = true;
+      component.view = 'full';
       fixture.detectChanges();
 
       for (const url of Object.keys(FAKE_DOCS)) {
@@ -126,9 +205,9 @@ describe('ExampleViewer', () => {
       button = btnDe ? btnDe.nativeElement : null;
     });
 
-    it('should call copier service when clicked', (() => {
-      const copierService: CopierService = TestBed.inject(CopierService);
-      const spy = spyOn(copierService, 'copyText');
+    it('should call clipboard service when clicked', (() => {
+      const clipboardService = TestBed.inject(Clipboard);
+      const spy = spyOn(clipboardService, 'copy');
       expect(spy.calls.count()).toBe(0, 'before click');
       button.click();
       expect(spy.calls.count()).toBe(1, 'after click');
@@ -137,18 +216,18 @@ describe('ExampleViewer', () => {
 
     it('should display a message when copy succeeds', (() => {
       const snackBar: MatSnackBar = TestBed.inject(MatSnackBar);
-      const copierService: CopierService = TestBed.inject(CopierService);
+      const clipboardService = TestBed.inject(Clipboard);
       spyOn(snackBar, 'open');
-      spyOn(copierService, 'copyText').and.returnValue(true);
+      spyOn(clipboardService, 'copy').and.returnValue(true);
       button.click();
       expect(snackBar.open).toHaveBeenCalledWith('Code copied', '', {duration: 2500});
     }));
 
     it('should display an error when copy fails', (() => {
       const snackBar: MatSnackBar = TestBed.inject(MatSnackBar);
-      const copierService: CopierService = TestBed.inject(CopierService);
+      const clipboardService = TestBed.inject(Clipboard);
       spyOn(snackBar, 'open');
-      spyOn(copierService, 'copyText').and.returnValue(false);
+      spyOn(clipboardService, 'copy').and.returnValue(false);
       button.click();
       expect(snackBar.open)
           .toHaveBeenCalledWith('Copy failed. Please try again!', '', {duration: 2500});
@@ -176,10 +255,8 @@ class TestExampleModule { }
 
 
 const FAKE_DOCS: {[key: string]: string} = {
-  '/docs-content/examples-highlighted/autocomplete-overview-example-html.html':
-      '<div>my docs page</div>',
-  '/docs-content/examples-highlighted/autocomplete-overview-example-ts.html':
-      '<span>const a = 1;</span>',
-  '/docs-content/examples-highlighted/autocomplete-overview-example-css.html':
+  [`${exampleBasePath}/autocomplete-overview-example-html.html`]: '<div>my docs page</div>',
+  [`${exampleBasePath}/autocomplete-overview-example-ts.html`]: '<span>const a = 1;</span>',
+  [`${exampleBasePath}/autocomplete-overview-example-css.html`]:
       '<pre>.class { color: black; }</pre>',
 };
