@@ -3,60 +3,93 @@ import {Injectable} from '@angular/core';
 export interface GuideItem {
   id: string;
   name: string;
+  nameCn: string;
   document: string;
+  overview: string;
+  overviewCn: string;
 }
 
 const GUIDES = [
   {
     id: 'getting-started',
-    name: '快速上手',
+    name: 'Getting started',
+    nameCn: '快速上手',
     document: '/docs-content/guides/getting-started.html',
+    overview: 'Add Angular Material to your project!',
+    overviewCn: '把 Angular Material 添加到你的项目中！',
   },
   {
     id: 'schematics',
-    name: '生成器（schematics）',
-    document: '/docs-content/guides/schematics.html'
+    name: 'Schematics',
+    nameCn: '原理图',
+    document: '/docs-content/guides/schematics.html',
+    overview: 'Use schematics to quickly generate views with Material Design components.',
+    overviewCn: '使用原理图快速生成带有 Material Design 组件的视图。',
   },
   {
     id: 'theming',
-    name: 'Angular Material 主题',
+    name: 'Theming Angular Material',
+    nameCn: 'Angular Material 主题',
     document: '/docs-content/guides/theming.html',
+    overview: `Customize your application with Angular Material's theming system.`,
+    overviewCn: '使用 Angular Material 的主题体系定制你的应用。',
   },
   {
     id: 'theming-your-components',
-    name: '为你自己的组件设置主题',
+    name: 'Theming your own components',
+    nameCn: '为你自己的组件设置主题',
     document: '/docs-content/guides/theming-your-components.html',
+    overview: `Use Angular Material's theming system in your own custom components.`,
+    overviewCn: '在你的自定义组件中使用 Angular Material 的主题体系。',
   },
   {
     id: 'typography',
-    name: `使用 Angular Material 的文字风格`,
+    name: `Customizing Typography`,
+    nameCn: '自定义排版',
     document: '/docs-content/guides/typography.html',
+    overview: 'Configure the typography settings for Angular Material components.',
+    overviewCn: '为 Angular Material 组件配置排版设置。',
   },
   {
     id: 'customizing-component-styles',
-    name: '自定义组件样式',
-    document: '/docs-content/guides/customizing-component-styles.html'
+    name: 'Customizing component styles',
+    nameCn: '自定义组件样式',
+    document: '/docs-content/guides/customizing-component-styles.html',
+    overview: 'Understand how to approach style customization with Angular Material components.',
+    overviewCn: '了解如何定制 Angular Material 组件的样式。',
   },
   {
     id: 'creating-a-custom-form-field-control',
-    name: '创建自定义表单控件',
-    document: '/docs-content/guides/creating-a-custom-form-field-control.html'
+    name: 'Custom form field control',
+    nameCn: '自定义表单控件',
+    document: '/docs-content/guides/creating-a-custom-form-field-control.html',
+    overview: 'Build a custom control that integrates with `<mat-form-field>`.',
+    overviewCn: '构建一个可以和 `<mat-form-field>` 集成的自定义表单控件。',
   },
   {
     id: 'elevation',
-    name: '使用 Z 轴助手',
-    document: '/docs-content/guides/elevation.html'
+    name: 'Elevation helpers',
+    nameCn: '纵深助手',
+    document: '/docs-content/guides/elevation.html',
+    overview: 'Enhance your components with elevation and depth.',
+    overviewCn: '利用纵深感强化你的组件。',
   },
   {
     id: 'creating-a-custom-stepper-using-the-cdk-stepper',
-    name: 'Creating a custom stepper using the CdkStepper',
-    document: '/docs-content/guides/creating-a-custom-stepper-using-the-cdk-stepper.html'
+    name: 'Custom stepper using the CdkStepper',
+    nameCn: '使用 CdkStepper 自定义步进器',
+    document: '/docs-content/guides/creating-a-custom-stepper-using-the-cdk-stepper.html',
+    overview: 'Create a custom stepper components using Angular CDK.',
+    overviewCn: '使用 Angular CDK 创建自定义步进器组件。',
   },
   {
     id: 'using-component-harnesses',
-    name: `Using Angular Material's component harnesses in your tests`,
-    document: '/docs-content/guides/using-component-harnesses.html'
-  }
+    name: `Testing with component harnesses`,
+    nameCn: '使用组件测试工具进行测试',
+    document: '/docs-content/guides/using-component-harnesses.html',
+    overview: 'Write tests with component harnesses for convenience and meaningful results.',
+    overviewCn: '使用组件测试工具编写测试，以获得便利性，并输出更有意义的结果。',
+  },
 ];
 
 @Injectable()
