@@ -1,4 +1,4 @@
-import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import {waitForAsync, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {NavBar, NavBarModule} from './navbar';
 import {DocsAppTestingModule} from '../../testing/testing-module';
@@ -7,7 +7,7 @@ import {DocsAppTestingModule} from '../../testing/testing-module';
 describe('NavBar', () => {
   let fixture: ComponentFixture<NavBar>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       imports: [NavBarModule, DocsAppTestingModule],
     }).compileComponents();
