@@ -1,4 +1,4 @@
-import {Component, NgModule} from '@angular/core';
+import {Component, NgModule, OnDestroy} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {MatButtonModule} from '@angular/material/button';
 import {MatMenuModule} from '@angular/material/menu';
@@ -9,6 +9,8 @@ import {SECTIONS} from '../documentation-items/documentation-items';
 import {ThemeStorage} from '../theme-picker/theme-storage/theme-storage';
 import {StyleManager} from '../style-manager';
 import {HttpClientModule} from '@angular/common/http';
+import {Subscription} from 'rxjs';
+import {NavigationFocusService} from '../navigation-focus/navigation-focus.service';
 
 const SECTIONS_KEYS = Object.keys(SECTIONS);
 
@@ -17,8 +19,15 @@ const SECTIONS_KEYS = Object.keys(SECTIONS);
   templateUrl: './navbar.html',
   styleUrls: ['./navbar.scss']
 })
-export class NavBar {
+export class NavBar implements OnDestroy {
+  private subscriptions = new Subscription();
   isNextVersion = location.hostname.startsWith('next.material.angular.io');
+  skipLinkHref: string|null;
+  skipLinkHidden = true;
+
+  constructor(private navigationFocusService: NavigationFocusService) {
+    setTimeout(() => this.skipLinkHref = this.navigationFocusService.getSkipLinkHref(), 100);
+  }
 
   get sections() {
     return SECTIONS;
@@ -26,6 +35,10 @@ export class NavBar {
 
   get sectionKeys() {
     return SECTIONS_KEYS;
+  }
+
+  ngOnDestroy() {
+    this.subscriptions.unsubscribe();
   }
 }
 
