@@ -1,10 +1,10 @@
-import {async, TestBed} from '@angular/core/testing';
+import {waitForAsync, TestBed} from '@angular/core/testing';
 import {ThemePicker, ThemePickerModule} from './theme-picker';
 import {DocsAppTestingModule} from '../../testing/testing-module';
 
 
 describe('ThemePicker', () => {
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       imports: [ThemePickerModule, DocsAppTestingModule],
     }).compileComponents();
